requirements.txt
pytest==8.3.5
selenium==4.30
allure-pytest==2.13.5
pytest-dotenv==0.5.2
PyHamcrest==2.1.0